// backend/src/API/types/AuthSchema.ts
import { z } from 'zod';

export const signupSchema = z.object({
    name: z.string().min(3, 'Name must be at least 3 characters long'),
    email: z.string().email('Invalid email format'),
    password: z.string().min(6, 'Password must be at least 6 characters long'),
});

export const signinSchema = z.object({
    email: z.string().email('Invalid email format'),
    password: z.string(),
});